+++ module/nps_wheres_wally/views/widget.edit.php
@@ -12,5 +12,5 @@
 
 (new CWidgetFormView($data))
     ->addField(new CWidgetFieldMultiSelectItemView($data['fields']['itemid']))
-    ->addField(new CWidgetFieldNumericBoxView($data['fields']['show_lines']))
+    ->addField(new CWidgetFieldIntegerBoxView($data['fields']['show_lines']))
     ->show();